requirements.txt
llama-cpp-python
fastapi
uvicorn
nest-asyncio
pyngrok